refactor: extract shared useDeleteCell hook to eliminate code duplication
- Create useDeleteCell hook with shared deletion logic
- Remove 82 lines of duplicate code between ExecutableCell and MarkdownCell
- Centralize focus management and delete logic in one place
- Maintain same functionality with better maintainability
