+++ src/components/notebook/cell/ExecutableCell.tsx
@@ -7,6 +7,7 @@ import { useAuth } from "@/components/auth/AuthProvider.js";
 import { useUserRegistry } from "@/hooks/useUserRegistry.js";
 import { useInterruptExecution } from "@/hooks/useInterruptExecution.js";
 import { useEditorRegistry } from "@/hooks/useEditorRegistry.js";
+import { useDeleteCell } from "@/hooks/useDeleteCell.js";
 
 import { useStore } from "@livestore/react";
 import { focusedCellSignal$, hasManuallyFocused$ } from "../signals/focus.js";
@@ -79,6 +80,8 @@ export const ExecutableCell: React.FC<ExecutableCellProps> = ({
     unregisterEditor,
     focusCell: registryFocusCell,
   } = useEditorRegistry();
+
+  const { handleDeleteCell } = useDeleteCell(cell.id);
 
   const {
     user: { sub: userId },
@@ -255,48 +258,6 @@ export const ExecutableCell: React.FC<ExecutableCellProps> = ({
       }
     },
     [cell.id, store, registryFocusCell]
-  );
-
-  // Cell-level delete handler
-  const handleDeleteCell = useCallback(
-    (trigger: "keyboard" | "click" = "click") => {
-      const cellReferences = store.query(queries.cellsWithIndices$);
-      const currentIndex = cellReferences.findIndex((c) => c.id === cell.id);
-
-      // Focus management based on trigger type
-      if (trigger === "keyboard") {
-        // Backspace: focus previous cell
-        if (currentIndex > 0) {
-          const previousCell = cellReferences[currentIndex - 1];
-          store.setSignal(focusedCellSignal$, previousCell.id);
-          setTimeout(() => {
-            registryFocusCell(previousCell.id, "end");
-          }, 0);
-        } else if (cellReferences.length > 1) {
-          // Focus next cell if deleting the first cell
-          const nextCell = cellReferences[currentIndex + 1];
-          store.setSignal(focusedCellSignal$, nextCell.id);
-          setTimeout(() => {
-            registryFocusCell(nextCell.id, "start");
-          }, 0);
-        } else {
-          // Last cell - clear focus
-          store.setSignal(focusedCellSignal$, null);
-        }
-      } else if (trigger === "click") {
-        // Click delete: clear focus
-        store.setSignal(focusedCellSignal$, null);
-      }
-
-      // Delete the cell
-      store.commit(
-        events.cellDeleted({
-          id: cell.id,
-          actorId: userId,
-        })
-      );
-    },
-    [store, cell.id, userId, registryFocusCell]
   );
 
   // Use shared keyboard navigation hook with cell-type-specific execution
