+++ src/components/notebook/cell/MarkdownCell.tsx
@@ -2,6 +2,7 @@ import { useCellContent } from "@/hooks/useCellContent.js";
 import { useCellKeyboardNavigation } from "@/hooks/useCellKeyboardNavigation.js";
 import { useCellOutputs } from "@/hooks/useCellOutputs.js";
 import { useEditorRegistry } from "@/hooks/useEditorRegistry.js";
+import { useDeleteCell } from "@/hooks/useDeleteCell.js";
 import { useStore } from "@livestore/react";
 import { events, tables, queries } from "@/schema";
 import React, {
@@ -58,6 +59,8 @@ export const MarkdownCell: React.FC<MarkdownCellProps> = ({
     unregisterEditor,
     focusCell: registryFocusCell,
   } = useEditorRegistry();
+
+  const { handleDeleteCell } = useDeleteCell(cell.id);
   // Use shared content management hook
   const { localSource, setLocalSource, updateSource, handleSourceChange } =
     useCellContent({
@@ -175,48 +178,6 @@ export const MarkdownCell: React.FC<MarkdownCellProps> = ({
       }
     },
     [cell.id, store, registryFocusCell]
-  );
-
-  // Cell-level delete handler
-  const handleDeleteCell = useCallback(
-    (trigger: "keyboard" | "click" = "click") => {
-      const cellReferences = store.query(queries.cellsWithIndices$);
-      const currentIndex = cellReferences.findIndex((c) => c.id === cell.id);
-
-      // Focus management based on trigger type
-      if (trigger === "keyboard") {
-        // Backspace: focus previous cell
-        if (currentIndex > 0) {
-          const previousCell = cellReferences[currentIndex - 1];
-          store.setSignal(focusedCellSignal$, previousCell.id);
-          setTimeout(() => {
-            registryFocusCell(previousCell.id, "end");
-          }, 0);
-        } else if (cellReferences.length > 1) {
-          // Focus next cell if deleting the first cell
-          const nextCell = cellReferences[currentIndex + 1];
-          store.setSignal(focusedCellSignal$, nextCell.id);
-          setTimeout(() => {
-            registryFocusCell(nextCell.id, "start");
-          }, 0);
-        } else {
-          // Last cell - clear focus
-          store.setSignal(focusedCellSignal$, null);
-        }
-      } else if (trigger === "click") {
-        // Click delete: clear focus
-        store.setSignal(focusedCellSignal$, null);
-      }
-
-      // Delete the cell
-      store.commit(
-        events.cellDeleted({
-          id: cell.id,
-          actorId: userId,
-        })
-      );
-    },
-    [store, cell.id, userId, registryFocusCell]
   );
 
   // Use shared keyboard navigation hook
